function/requirements.txt
boto3==1.42.94
botocore==1.42.94
cffi==2.0.0
cryptography==46.0.7
jmespath==1.1.0
pycparser==3.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
s3transfer==0.16.1
six==1.17.0
urllib3==2.6.3
